Block List: Move BlockList to layout file
Aiding in GitHub rebase history as explicit move
